Removed duplicate IDs, wrong values, and dev stuff
Removed developer headers in index.php

Removed 'duplicate' HTML ids in generation of forms. It was creating
empty strings for ids and was causing errors in Google Chrome.

Removed the second POC having his/her email showing up in a CAP ID list,
when their CAP ID was supposed to be showing up